refactor(test): split SQLite executor tests
